+++ src/app/articles/page.tsx
@@ -4,7 +4,7 @@ import Newsletter from "../components/Newsletter";
 import TrendingArticles from "../components/TrendingArticles";
 import AdSlot from "../components/AdSlot";
 import { AD_SLOTS } from "@/lib/constants";
-import { fetchArticles } from "@/lib/articles";
+import { fetchArticles } from "@/lib/articles.server";
 
 export const metadata = {
   title: "All Articles",
