+++ src/lib/mongodb.ts
@@ -1,3 +1,5 @@
+import "server-only";
+
 import { MongoClient } from "mongodb";
 
 const uri = process.env.MONGODB_URI;
